+++ CropGrowthInfo/TextData.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.Xna.Framework;
+
+namespace CropGrowthInfo
+{
+	public class TextData
+	{
+		public Color color;
+		public string text;
+
+		public TextData(string _text, Color _color)
+		{
+			text = _text;
+			color = _color;
+		}
+	}
+}
