--- CropGrowthInformation/i18n/fr.json
@@ -1,30 +1,0 @@
-{
-  "name": "Nom de la culture",
-  "phase": "Étape {0}/{1}",
-  "current": "Jour {0}/{1}",
-  "total": "Croissance {0}/{1}",
-  "ready": "Prête !",
-
-  // GMCM
-  "GMCM_Option_ModEnabled_Name": "Activer le Mod",
-  "GMCM_Title_Keys_Text": "Touches",
-  "GMCM_Option_RequireToggle_Name": "Nécessite la touche d'activation",
-  "GMCM_Option_ToggleButton_Name": "Touche d'activation",
-  "GMCM_Title_Information_Text": "Informations",
-  "GMCM_Option_ShowCropName_Name": "Afficher le nom de la culture",
-  "GMCM_Option_ShowCurrentPhase_Name": "Afficher l'étape actuelle",
-  "GMCM_Option_ShowDaysInCurrentPhase_Name": "Afficher les jours de l'étape",
-  "GMCM_Option_ShowTotalGrowth_Name": "Afficher le nombre total de jours",
-  "GMCM_Option_ShowReadyText_Name": "Afficher un texte lorsque prête",
-  "GMCM_Title_Display_Text": "Affichage",
-  "GMCM_Option_TextScale_Name": "Échelle du texte",
-  "GMCM_Option_CropNameColor_Name": "Couleur du nom de la culture",
-  "GMCM_Option_CurrentPhaseColor_Name": "Couleur de l'étape actuelle",
-  "GMCM_Option_DaysInCurrentPhaseColor_Name": "Couleur des jours de l'étape",
-  "GMCM_Option_TotalGrowthColor_Name": "Couleur du nombre total de jours",
-  "GMCM_Option_ReadyTextColor_Name": "Couleur du texte lorsque prête",
-  "GMCM_Option_Red_Name": "Rouge",
-  "GMCM_Option_Green_Name": "Vert",
-  "GMCM_Option_Blue_Name": "Bleu",
-  "GMCM_Option_CropOpacity_Name": "Opacité de la culture"
-}
